:truck: (interface): Rename file interface/CoreLTDC.hpp to interface/LTDC.hpp
